Move validators.TestManager to validatorstest.Manager
