+++ vms/platformvm/validators/validatorstest/manager.go
@@ -1,16 +1,18 @@
 // Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package validators
+package validatorstest
 
 import (
 	"context"
 
 	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/snow/validators"
+
+	snowvalidators "github.com/ava-labs/avalanchego/snow/validators"
+	vmvalidators "github.com/ava-labs/avalanchego/vms/platformvm/validators"
 )
 
-var TestManager Manager = testManager{}
+var Manager vmvalidators.Manager = testManager{}
 
 type testManager struct{}
 
@@ -26,7 +28,7 @@ func (testManager) GetSubnetID(context.Context, ids.ID) (ids.ID, error) {
 	return ids.Empty, nil
 }
 
-func (testManager) GetValidatorSet(context.Context, uint64, ids.ID) (map[ids.NodeID]*validators.GetValidatorOutput, error) {
+func (testManager) GetValidatorSet(context.Context, uint64, ids.ID) (map[ids.NodeID]*snowvalidators.GetValidatorOutput, error) {
 	return nil, nil
 }
